# server/docs/swagger.yaml
definitions:
...
  request.SysAuthorityBtnReq:
    properties:
      authorityId:
        type: integer
      menuID:
        type: integer
      selected:
        items:
          type: integer
        type: array
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  response.SysAuthorityBtnRes:
    properties:
      selected:
        items:
          type: integer
        type: array
    type: object
...
  system.AutoCodeStruct:
    properties:
      abbreviation:
        description: Struct简称
        type: string
      autoCreateApiToSql:
        description: 是否自动创建api
        type: boolean
      autoCreateMenuToSql:
        description: 是否自动创建menu
        type: boolean
      autoCreateResource:
        description: 是否自动创建资源标识
        type: boolean
      businessDB:
        description: 业务数据库
        type: string
      description:
        description: Struct中文名称
        type: string
      fields:
        items:
          $ref: '#/definitions/system.Field'
        type: array
      gvaModel:
        description: 是否使用gva默认Model
        type: boolean
      humpPackageName:
        description: go文件名称
        type: string
      package:
        type: string
      packageName:
        description: 文件名称
        type: string
      primaryField:
        $ref: '#/definitions/system.Field'
      structName:
        description: Struct名称
        type: string
      tableName:
        description: 表名
        type: string
    type: object
...
  system.DataSource:
    properties:
      association:
        description: 关联关系 1 一对一 2 一对多
        type: integer
      label:
        type: string
      table:
        type: string
      value:
        type: string
    type: object
  system.Field:
    properties:
      checkDataSource:
        description: 是否检查数据源
        type: boolean
      clearable:
        description: 是否可清空
        type: boolean
      columnName:
        description: 数据库字段
        type: string
      comment:
        description: 数据库字段描述
        type: string
      dataSource:
        allOf:
        - $ref: '#/definitions/system.DataSource'
        description: 数据源
      dataTypeLong:
        description: 数据库字段长度
        type: string
      defaultValue:
        description: 是否必填
        type: string
      dictType:
        description: 字典
        type: string
      errorText:
        description: 校验失败文字
        type: string
      fieldDesc:
        description: 中文名
        type: string
      fieldJson:
        description: FieldJson
        type: string
      fieldName:
        description: Field名
        type: string
      fieldSearchHide:
        description: 是否隐藏查询条件
        type: boolean
      fieldSearchType:
        description: 搜索条件
        type: string
      fieldType:
        description: Field数据类型
        type: string
      front:
        description: 是否前端可见
        type: boolean
      primaryKey:
        description: 是否主键
        type: boolean
      require:
        description: 是否必填
        type: boolean
      sort:
        description: 是否增加排序
        type: boolean
    type: object
...
  system.SysAutoCode:
    properties:
      ID:
        description: 主键ID
        type: integer
      createdAt:
        description: 创建时间
        type: string
      desc:
        type: string
      label:
        type: string
      packageName:
        type: string
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.6.5
paths:
...
  /authorityBtn/canRemoveAuthorityBtn:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 设置权限按钮
      tags:
      - AuthorityBtn
  /authorityBtn/getAuthorityBtn:
    post:
      consumes:
      - application/json
      parameters:
      - description: 菜单id, 角色id, 选中的按钮id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.SysAuthorityBtnReq'
      produces:
      - application/json
      responses:
        "200":
          description: 返回列表成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.SysAuthorityBtnRes'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取权限按钮
      tags:
      - AuthorityBtn
  /authorityBtn/setAuthorityBtn:
    post:
      consumes:
      - application/json
      parameters:
      - description: 菜单id, 角色id, 选中的按钮id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.SysAuthorityBtnReq'
      produces:
      - application/json
      responses:
        "200":
          description: 返回列表成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 设置权限按钮
      tags:
      - AuthorityBtn
  /autoCode/createPackage:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建package
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.SysAutoCode'
      produces:
      - application/json
      responses:
        "200":
          description: 创建package成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建package
      tags:
      - AutoCode
  /autoCode/createPlug:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建插件模板
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.SysAutoCode'
      produces:
      - application/json
      responses:
        "200":
          description: 创建插件模板成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建插件模板
      tags:
      - AutoCode
  /autoCode/createTemp:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建自动代码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.AutoCodeStruct'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"创建成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 自动代码模板
      tags:
      - AutoCode
  /autoCode/delPackage:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建package
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.SysAutoCode'
      produces:
      - application/json
      responses:
        "200":
          description: 删除package成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除package
      tags:
      - AutoCode
...
  /autoCode/getColumn:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 获取当前表所有字段
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取当前表所有字段
      tags:
      - AutoCode
  /autoCode/getDatabase:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 获取当前所有数据库
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取当前所有数据库
      tags:
      - AutoCode
...
  /autoCode/getPackage:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 创建package成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取package
      tags:
      - AutoCode
...
  /autoCode/getTables:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 获取当前数据库所有表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取当前数据库所有表
      tags:
      - AutoCode
  /autoCode/installPlugin:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: this is a test file
        in: formData
        name: plug
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 安装插件成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    type: object
                  type: array
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 安装插件
      tags:
      - AutoCode
  /autoCode/preview:
    post:
      consumes:
      - application/json
      parameters:
      - description: 预览创建代码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.AutoCodeStruct'
      produces:
      - application/json
      responses:
        "200":
          description: 预览创建后的代码
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 预览创建后的代码
      tags:
      - AutoCode
  /autoCode/pubPlug:
    get:
      consumes:
      - application/json
      parameters:
      - description: 打包插件
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.SysAutoCode'
      produces:
      - application/json
      responses:
        "200":
          description: 打包插件成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 打包插件
      tags:
      - AutoCode
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"